requirements.txt
watchdog>=0.8.3
PyNMA>=1.0
python-telegram-bot>=4.1.2
PyYAML>=3.11

